# Improve performance of join building performance (up to 1.28x speedup)